pyproject.toml
[project]
name = "trading-rl-env"
version = "0.1.0"
description = "RL environment for simulated stock trading using QuantReplay"
requires-python = ">=3.11"
dependencies = [
    "hud-python[agents]>=0.5.17",
    "httpx>=0.24.0",
    "simplefix>=1.2",
    "numpy>=1.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hud]
image = "trading-rl-env:dev"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "B", "I", "C4", "UP"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["backend", "grading", "tasks", "tools"]
